+++ libft/libft.h
@@ -48,8 +48,7 @@ int		ft_tolower(int c);
 int		ft_strncmp(const char *s1, const char *s2, size_t n);
 int		ft_atoi(const char *nptr);
 
-
-char	**free_split(char **split, int index);
+char	**free_split(char **split, ssize_t index);
 char	*ft_itoa(int n);
 char	*ft_strchr(const char *s, int c);
 char	*ft_strrchr(const char *s, int c);
